# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Play9ja Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/otp/send:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password/reset-request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/password/reset:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordReset"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /deposit:
    get:
      operationId: listDeposits
      tags: [deposit]
      summary: List user deposits
      responses:
        "200":
          description: Deposits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [deposit]
      summary: Submit a deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposit/{id}:
    get:
      operationId: getDeposit
      tags: [deposit]
      summary: Get a single deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"

  # WITHDRAWAL
...
  /games/{id}/play:
    post:
      operationId: playGame
      tags: [games]
      summary: Play a game and get result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GamePlayInput"
      responses:
        "200":
          description: Game result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameResult"
        "403":
          description: Daily limit reached or premium required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /games/{id}/history:
    get:
      operationId: getGameHistory
      tags: [games]
      summary: Get history for a specific game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GamePlay"

  # LEADERBOARD
...
  /admin/deposits/{id}:
    patch:
      operationId: updateDepositStatus
      tags: [admin]
      summary: Approve or decline a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"

  # ADMIN - WITHDRAWALS
...
  /admin/games/{id}:
    patch:
      operationId: updateGame
      tags: [admin]
      summary: Update game settings (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isSuspended:
          type: boolean
        referralCode:
          type: string
        createdAt:
          type: string
          format: date-time
...
    UserRegistrationInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        referralCode:
          type: ["string", "null"]
        rememberMe:
          type: boolean
    UserLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        isNewUser:
          type: boolean
        welcomeBonus:
          type: ["number", "null"]
    OtpRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    OtpVerify:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    PasswordReset:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
...
    Deposit:
      type: object
      required: [id, userId, amount, status, accountName, accountNumber, bankName, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, declined]
        accountName:
          type: string
        accountNumber:
          type: string
        bankName:
          type: string
        proofUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DepositInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 100
        proofUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Game:
      type: object
      required: [id, name, description, emoji, category, isEnabled, dailyLimit, minReward, maxReward, isPremium]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        emoji:
          type: string
        category:
          type: string
        isEnabled:
          type: boolean
        dailyLimit:
          type: integer
        minReward:
          type: number
        maxReward:
          type: number
        isPremium:
          type: boolean
        premiumMultiplier:
          type: number
        difficulty:
          type: string
          enum: [easy, medium, hard]
        playsToday:
          type: integer
        totalPlays:
          type: integer
    GamePlayInput:
      type: object
      properties:
        bet:
          type: ["number", "null"]
        choice:
          type: ["string", "null"]
    GameResult:
      type: object
      required: [won, reward, message, gamePlayId]
      properties:
        won:
          type: boolean
        reward:
          type: number
        message:
          type: string
        gamePlayId:
          type: integer
        animation:
          type: string
    GamePlay:
      type: object
      required: [id, gameId, gameName, won, reward, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        gameName:
          type: string
        won:
          type: boolean
        reward:
          type: number
        createdAt:
          type: string
          format: date-time
...
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, declined]
        notes:
          type: ["string", "null"]
...
    GameUpdate:
      type: object
      properties:
        isEnabled:
          type: boolean
        dailyLimit:
          type: integer
        minReward:
          type: number
        maxReward:
          type: number
        isPremium:
          type: boolean
        premiumMultiplier:
          type: number